turtle-bot-main/requirements.txt
aiogram==2.20
python-dotenv==0.20.0
emoji~=1.7.0
gino~=1.0.1
SQLAlchemy~=1.3.24
asyncpg~=0.25.0